requirements.txt
streamlit>=1.38
pandas>=2.2
numpy>=1.26
plotly>=5.24
matplotlib>=3.9
scipy>=1.13
pydantic>=2.9
python-docx>=1.1
reportlab>=4.2
kaleido>=0.2.1
